# internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A minimal, self-hosted OpenAI-compatible AI gateway: model-name routing, retries/failover, per-key limits, token accounting, idempotent replay, and OpenTelemetry.\nBodies are OpenAI-compatible and forwarded verbatim; mortise only inspects model, stream, and usage.",
        "title": "mortise",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
        "/healthz": {
            "get": {
                "description": "Always returns 200 while the server is running. No auth required.",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Liveness probe",
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/chat/completions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Routes to the pool serving `model`, applying per-key rate/token limits, retries/failover, and optional idempotent replay. Supports SSE when `stream` is true.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json",
                    "text/event-stream"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Create a chat completion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Makes the request idempotent; retries replay the original response",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "OpenAI-compatible chat request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.ChatCompletionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ChatCompletionResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/models": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lists the client-visible model names mortise is configured to route.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Models"
                ],
                "summary": "List available models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ModelList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.ChatCompletionRequest": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.ChatMessage"
                    }
                },
                "model": {
                    "type": "string",
                    "example": "llama-3.1-8b"
                },
                "stream": {
                    "type": "boolean"
                },
                "stream_options": {
                    "$ref": "#/definitions/server.StreamOptions"
                }
            }
        },
        "server.ChatCompletionResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "chatcmpl-abc123"
                },
                "model": {
                    "type": "string",
                    "example": "llama-3.1-8b"
                },
                "object": {
                    "type": "string",
                    "example": "chat.completion"
                },
                "usage": {
                    "$ref": "#/definitions/server.UsageBlock"
                }
            }
        },
        "server.ChatMessage": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "example": "Hello!"
                },
                "role": {
                    "type": "string",
                    "enum": [
                        "system",
                        "user",
                        "assistant",
                        "tool"
                    ],
                    "example": "user"
                }
            }
        },
        "server.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "invalid_api_key"
                },
                "message": {
                    "type": "string",
                    "example": "invalid api key"
                },
                "type": {
                    "type": "string",
                    "example": "invalid_request_error"
                }
            }
        },
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/server.ErrorDetail"
                }
            }
        },
        "server.ModelDoc": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "llama-3.1-8b"
                },
                "object": {
                    "type": "string",
                    "example": "model"
                },
                "owned_by": {
                    "type": "string",
                    "example": "mortise"
                }
            }
        },
        "server.ModelList": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.ModelDoc"
                    }
                },
                "object": {
                    "type": "string",
                    "example": "list"
                }
            }
        },
        "server.StreamOptions": {
            "type": "object",
            "properties": {
                "include_usage": {
                    "description": "IncludeUsage emits a final chunk carrying usage (needed for streaming\ntoken accounting).",
                    "type": "boolean"
                }
            }
        },
        "server.UsageBlock": {
            "type": "object",
            "properties": {
                "completion_tokens": {
                    "type": "integer"
                },
                "prompt_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Client API key presented as \"Bearer \u003ckey\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}